Add support for NavigationTypeProperty in V4 Functions and elements removed/replaced in V4 corresponding with NavigationTypeProperty were moved to v2 module.
