+++ pyodata/v2/__init__.py
@@ -9,9 +9,12 @@ from pyodata.model.type_traits import EdmBooleanTypTraits, EdmPrefixedTypTraits,
     EdmLongIntTypTraits, EdmStringTypTraits
 from pyodata.config import ODATAVersion
 
+from pyodata.v2.elements import NavigationTypeProperty, EndRole, Association, AssociationSetEndRole, AssociationSet, \
+    ReferentialConstraint, Schema
 from pyodata.model.elements import StructTypeProperty, StructType, ComplexType, EntityType, \
-    EnumType, EntitySet, ReferentialConstraint, ExternalAnnotation, Annotation, ValueHelper, ValueHelperParameter, \
-    FunctionImport, Schema, Typ, NavigationTypeProperty, EndRole, Association, AssociationSetEndRole, AssociationSet
+    EnumType, EntitySet, ExternalAnnotation, Annotation, ValueHelper, ValueHelperParameter, \
+    FunctionImport, Typ
+
 
 import pyodata.v2.build_functions as build_functions_v2
 import pyodata.model.build_functions as build_functions
@@ -35,11 +38,11 @@ class ODataV2(ODATAVersion):
             EntityType: build_functions.build_entity_type,
             EnumType: build_functions.build_enum_type,
             EntitySet: build_functions.build_entity_set,
-            EndRole: build_functions.build_end_role,
-            ReferentialConstraint: build_functions.build_referential_constraint,
-            Association: build_functions.build_association,
-            AssociationSetEndRole: build_functions.build_association_set_end_role,
-            AssociationSet: build_functions.build_association_set,
+            EndRole: build_functions_v2.build_end_role,
+            ReferentialConstraint: build_functions_v2.build_referential_constraint,
+            Association: build_functions_v2.build_association,
+            AssociationSetEndRole: build_functions_v2.build_association_set_end_role,
+            AssociationSet: build_functions_v2.build_association_set,
             ExternalAnnotation: build_functions.build_external_annotation,
             Annotation: build_functions.build_annotation,
             ValueHelper: build_functions.build_value_helper,
