+++ pyodata/v4/__init__.py
@@ -4,10 +4,11 @@ from typing import List
 
 from pyodata.config import ODATAVersion
 from pyodata.model.type_traits import EdmBooleanTypTraits, EdmIntTypTraits
-from pyodata.model.elements import Typ, Schema, EnumType, ComplexType, StructType, StructTypeProperty
+from pyodata.model.elements import Typ, Schema, EnumType, ComplexType, StructType, StructTypeProperty, EntityType
+
+from pyodata.v4.elements import NavigationTypeProperty
 from pyodata.v4.type_traits import EdmDateTypTraits, GeoTypeTraits, EdmDoubleQuotesEncapsulatedTypTraits, \
     EdmTimeOfDay, EdmDateTimeOffsetTypTraits, EdmDuration
-
 
 import pyodata.v4.build_functions as build_functions_v4
 import pyodata.model.build_functions as build_functions
@@ -21,9 +22,10 @@ class ODataV4(ODATAVersion):
         return {
             StructTypeProperty: build_functions.build_struct_type_property,
             StructType: build_functions.build_struct_type,
-            # NavigationTypeProperty: navigation_type_property_from_etree,
+            NavigationTypeProperty: build_functions_v4.build_navigation_type_property,
             EnumType: build_functions.build_enum_type,
             ComplexType: build_functions.build_complex_type,
+            EntityType: build_functions.build_entity_type,
             Schema: build_functions_v4.build_schema,
         }
